refactor: clean up agent top level
